git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/org.objectweb.proactive.ic2d.JMXmonitoring.ic3d@9083 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/org/objectweb/proactive/ic2d/jmxmonitoring/ic3d/views/ActiveObject3D.java
@@ -59,7 +59,7 @@ public class ActiveObject3D extends AbstractActiveObject3D {
 		Vector3d oldScale = new Vector3d();
 		resize.getScale(oldScale);
 		//TODO remove constant
-		resize.setScale(new Vector3d(size / 2, oldScale.y, oldScale.z));
+		resize.setScale(new Vector3d((double)size / 2, oldScale.y, oldScale.z));
 		trans.setTransform(resize);
 	}
